refactor run_sync control flow
